requirements.txt
pytest==6.2.1
requests==2.25.1
vcrpy==4.1.1s